[ci] Split pipeline e2e into stages
