+++ .github/workflows/flink_cdc_ci.yml
@@ -100,32 +100,22 @@ jobs:
       custom-maven-parameter: "-Pflink2"
       modules: "['mysql-source', 'postgres-source, oceanbase-source, tidb, vitess', 'oracle, sqlserver', 'db2, mongodb']"
   pipeline_e2e:
-    strategy:
-      max-parallel: 2
-      fail-fast: false
-      matrix:
-        parallelism: [ 1, 4 ]
-    name: Pipeline E2E Tests (${{ matrix.parallelism }}-Parallelism)
+    name: Pipeline E2E Tests
     uses: ./.github/workflows/flink_cdc_base.yml
     with:
       java-versions: "[11]"
       flink-versions: "['1.20.3']"
       modules: "['pipeline_e2e']"
-      parallelism: ${{ matrix.parallelism }}
+      e2e-stages: "['stage1', 'stage2', 'stage3']"
   pipeline_e2e_2_x:
-    strategy:
-      max-parallel: 2
-      fail-fast: false
-      matrix:
-        parallelism: [ 1, 4 ]
-    name: Pipeline E2E Tests 2.x (${{ matrix.parallelism }}-Parallelism)
+    name: Pipeline E2E Tests 2.x
     uses: ./.github/workflows/flink_cdc_base.yml
     with:
       java-versions: "[11]"
       flink-versions: "['2.2.0']"
       custom-maven-parameter: "-Pflink2"
       modules: "['pipeline_e2e']"
-      parallelism: ${{ matrix.parallelism }}
+      e2e-stages: "['stage1', 'stage2', 'stage3']"
   source_e2e:
     name: Source E2E Tests
     uses: ./.github/workflows/flink_cdc_base.yml
