+++ .github/workflows/flink_cdc_ci_nightly.yml
@@ -97,33 +97,23 @@ jobs:
       modules: "['mysql-source', 'postgres-source, oceanbase-source, tidb, vitess', 'oracle, sqlserver', 'db2, mongodb']"
   pipeline_e2e:
     if: github.repository == 'apache/flink-cdc'
-    strategy:
-      max-parallel: 2
-      fail-fast: false
-      matrix:
-        parallelism: [ 1, 4 ]
-    name: Pipeline E2E Tests (${{ matrix.parallelism }}-Parallelism)
+    name: Pipeline E2E Tests
     uses: ./.github/workflows/flink_cdc_base.yml
     with:
       java-versions: "[17]"
       flink-versions: "['1.20.3']"
       modules: "['pipeline_e2e']"
-      parallelism: ${{ matrix.parallelism }}
+      e2e-stages: "['stage1', 'stage2', 'stage3']"
   pipeline_e2e_2_x:
     if: github.repository == 'apache/flink-cdc'
-    strategy:
-      max-parallel: 2
-      fail-fast: false
-      matrix:
-        parallelism: [ 1, 4 ]
-    name: Pipeline E2E Tests 2.x (${{ matrix.parallelism }}-Parallelism)
+    name: Pipeline E2E Tests 2.x
     uses: ./.github/workflows/flink_cdc_base.yml
     with:
       java-versions: "[17]"
       flink-versions: "['2.2.0']"
       custom-maven-parameter: "-Pflink2"
       modules: "['pipeline_e2e']"
-      parallelism: ${{ matrix.parallelism }}
+      e2e-stages: "['stage1', 'stage2', 'stage3']"
   source_e2e:
     if: github.repository == 'apache/flink-cdc'
     name: Source E2E Tests
