+++ flink-cdc-e2e-tests/flink-cdc-pipeline-e2e-tests/src/test/java/org/apache/flink/cdc/pipeline/tests/stage3/OracleE2eITCase.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.cdc.pipeline.tests;
+package org.apache.flink.cdc.pipeline.tests.stage3;
 
 import org.apache.flink.cdc.common.test.utils.TestUtils;
 import org.apache.flink.cdc.connectors.oracle.source.OracleSourceTestBase;
@@ -27,6 +27,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedClass;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.testcontainers.DockerClientFactory;
@@ -61,10 +63,17 @@ import static org.apache.flink.cdc.connectors.oracle.source.OracleSourceTestBase
 import static org.apache.flink.cdc.connectors.oracle.source.OracleSourceTestBase.TOP_USER;
 
 /** End-to-end tests for Oracle cdc pipeline job. */
+@ParameterizedClass
+@ValueSource(ints = {1, 4})
 public class OracleE2eITCase extends PipelineTestEnvironment {
     private static final Logger LOG = LoggerFactory.getLogger(OracleE2eITCase.class);
     protected static final Pattern COMMENT_PATTERN = Pattern.compile("^(.*)--.*$");
     private static OracleContainer oracle;
+
+    public OracleE2eITCase(int parallelism) {
+        super(parallelism);
+    }
+
     // ------------------------------------------------------------------------------------------
     // Oracle Variables (we always use Oracle as the data source for easier verifying)
     // ------------------------------------------------------------------------------------------
