+++ flink-cdc-e2e-tests/flink-cdc-pipeline-e2e-tests/src/test/java/org/apache/flink/cdc/pipeline/tests/stage3/SchemaEvolvingTransformE2eITCase.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.flink.cdc.pipeline.tests;
+package org.apache.flink.cdc.pipeline.tests.stage3;
 
 import org.apache.flink.cdc.connectors.mysql.testutils.UniqueDatabase;
 import org.apache.flink.cdc.pipeline.tests.utils.PipelineTestEnvironment;
@@ -24,6 +24,8 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedClass;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -36,11 +38,18 @@ import java.util.Collections;
 import java.util.List;
 
 /** E2e tests for Schema Evolution cases. */
-class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
-    private static final Logger LOG = LoggerFactory.getLogger(SchemaEvolveE2eITCase.class);
+@ParameterizedClass
+@ValueSource(ints = {1, 4})
+class SchemaEvolvingTransformE2eITCase extends PipelineTestEnvironment {
+    private static final Logger LOG =
+            LoggerFactory.getLogger(SchemaEvolvingTransformE2eITCase.class);
 
     protected final UniqueDatabase schemaEvolveDatabase =
             new UniqueDatabase(MYSQL, "schema_evolve", MYSQL_TEST_USER, MYSQL_TEST_PASSWORD);
+
+    SchemaEvolvingTransformE2eITCase(int parallelism) {
+        super(parallelism);
+    }
 
     @BeforeEach
     public void before() throws Exception {
@@ -63,14 +72,13 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 false,
                 Arrays.asList(
                         "AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=age}]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012, Eve, 17, 0], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012 -> Eve, 1012, Eve, 17, 0, 1024144, age < 20], op=INSERT, meta=()}",
                         "AlterColumnTypeEvent{tableId=%s.members, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}",
-                        "RenameColumnEvent{tableId=%s.members, nameMapping={age=precise_age}}",
                         "RenameColumnEvent{tableId=%s.members, nameMapping={gender=biological_sex}}",
                         "DropColumnEvent{tableId=%s.members, droppedColumnNames=[biological_sex]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013, Fiona, 16.0], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013 -> Fiona, 1013, Fiona, 16.0, 1026169, age < 20], op=INSERT, meta=()}",
                         "TruncateTableEvent{tableId=%s.members}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014, Gem, 17.0], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014 -> Gem, 1014, Gem, 17.0, 1028196, age < 20], op=INSERT, meta=()}",
                         "DropTableEvent{tableId=%s.members}"));
     }
 
@@ -82,13 +90,12 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 false,
                 false,
                 Arrays.asList(
-                        "AddColumnEvent{tableId=%s.merged, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=age}]}",
-                        "DataChangeEvent{tableId=%s.merged, before=[], after=[1012, Eve, 17, 0], op=INSERT, meta=()}",
+                        "AddColumnEvent{tableId=%s.merged, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=tag}]}",
+                        "DataChangeEvent{tableId=%s.merged, before=[], after=[1012 -> Eve, 1012, Eve, 17, 1024144, age < 20, 0], op=INSERT, meta=()}",
                         "AlterColumnTypeEvent{tableId=%s.merged, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}",
-                        "AddColumnEvent{tableId=%s.merged, addedColumns=[ColumnWithPosition{column=`precise_age` DOUBLE, position=AFTER, existedColumnName=gender}]}",
-                        "AddColumnEvent{tableId=%s.merged, addedColumns=[ColumnWithPosition{column=`biological_sex` TINYINT, position=AFTER, existedColumnName=precise_age}]}",
-                        "DataChangeEvent{tableId=%s.merged, before=[], after=[1013, Fiona, null, null, 16.0, null], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.merged, before=[], after=[1014, Gem, null, null, 17.0, null], op=INSERT, meta=()}"));
+                        "AddColumnEvent{tableId=%s.merged, addedColumns=[ColumnWithPosition{column=`biological_sex` TINYINT, position=AFTER, existedColumnName=gender}]}",
+                        "DataChangeEvent{tableId=%s.merged, before=[], after=[1013 -> Fiona, 1013, Fiona, 16.0, 1026169, age < 20, null, null], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.merged, before=[], after=[1014 -> Gem, 1014, Gem, 17.0, 1028196, age < 20, null, null], op=INSERT, meta=()}"));
     }
 
     @Test
@@ -100,7 +107,8 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 false,
                 Collections.emptyList(),
                 Arrays.asList(
-                        "Caused by: org.apache.flink.util.FlinkRuntimeException: Failed to apply schema change event.",
+                        "An exception was triggered from Schema change applying task. Job will fail now.",
+                        "org.apache.flink.util.FlinkRuntimeException: Failed to apply schema change event.",
                         "Caused by: org.apache.flink.cdc.common.exceptions.UnsupportedSchemaChangeEventException",
                         "org.apache.flink.runtime.JobException: Recovery is suppressed by NoRestartBackoffTimeStrategy"));
     }
@@ -114,13 +122,12 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 false,
                 Arrays.asList(
                         // Add column never succeeded, so age column will not appear.
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012, Eve, 17], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013, Fiona, null], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014, Gem, null], op=INSERT, meta=()}"),
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012 -> Eve, 1012, Eve, 17, 1024144, age < 20], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013 -> Fiona, 1013, Fiona, null, 1026169, age < 20], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014 -> Gem, 1014, Gem, null, 1028196, age < 20], op=INSERT, meta=()}"),
                 Arrays.asList(
                         "Failed to apply schema change AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=age}]}, but keeps running in tolerant mode. Caused by: UnsupportedSchemaChangeEventException{applyingEvent=AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=age}]}, exceptionMessage='Rejected schema change event since error.on.schema.change is enabled.', cause='null'}",
                         "Failed to apply schema change AlterColumnTypeEvent{tableId=%s.members, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}, but keeps running in tolerant mode. Caused by: UnsupportedSchemaChangeEventException{applyingEvent=AlterColumnTypeEvent{tableId=%s.members, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}, exceptionMessage='Rejected schema change event since error.on.schema.change is enabled.', cause='null'}",
-                        "Failed to apply schema change RenameColumnEvent{tableId=%s.members, nameMapping={age=precise_age}}, but keeps running in tolerant mode. Caused by: UnsupportedSchemaChangeEventException{applyingEvent=RenameColumnEvent{tableId=%s.members, nameMapping={age=precise_age}}, exceptionMessage='Rejected schema change event since error.on.schema.change is enabled.', cause='null'}",
                         "Failed to apply schema change RenameColumnEvent{tableId=%s.members, nameMapping={gender=biological_sex}}, but keeps running in tolerant mode. Caused by: UnsupportedSchemaChangeEventException{applyingEvent=RenameColumnEvent{tableId=%s.members, nameMapping={gender=biological_sex}}, exceptionMessage='Rejected schema change event since error.on.schema.change is enabled.', cause='null'}",
                         "Failed to apply schema change DropColumnEvent{tableId=%s.members, droppedColumnNames=[biological_sex]}, but keeps running in tolerant mode. Caused by: UnsupportedSchemaChangeEventException{applyingEvent=DropColumnEvent{tableId=%s.members, droppedColumnNames=[biological_sex]}, exceptionMessage='Rejected schema change event since error.on.schema.change is enabled.', cause='null'}",
                         "Failed to apply schema change TruncateTableEvent{tableId=%s.members}, but keeps running in tolerant mode. Caused by: UnsupportedSchemaChangeEventException{applyingEvent=TruncateTableEvent{tableId=%s.members}, exceptionMessage='Rejected schema change event since error.on.schema.change is enabled.', cause='null'}",
@@ -136,9 +143,9 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 false,
                 false,
                 Arrays.asList(
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012, Eve, 17], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013, Fiona, null], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014, Gem, null], op=INSERT, meta=()}"));
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012 -> Eve, 1012, Eve, 17, 1024144, age < 20], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013 -> Fiona, 1013, Fiona, null, 1026169, age < 20], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014 -> Gem, 1014, Gem, null, 1028196, age < 20], op=INSERT, meta=()}"));
     }
 
     @Test
@@ -151,12 +158,11 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 Collections.emptyList(),
                 Arrays.asList(
                         "An exception was triggered from Schema change applying task. Job will fail now.",
-                        "org.apache.flink.runtime.JobException: Recovery is suppressed by NoRestartBackoffTimeStrategy"));
+                        "org.apache.flink.util.FlinkRuntimeException: Failed to apply schema change event."));
     }
 
     @Test
     void testLenientSchemaEvolution() throws Exception {
-
         testGenericSchemaEvolution(
                 "lenient",
                 false,
@@ -164,18 +170,11 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 false,
                 Arrays.asList(
                         "AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=LAST, existedColumnName=null}]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012, Eve, 17, 0], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012 -> Eve, 1012, Eve, 17, 1024144, age < 20, 0], op=INSERT, meta=()}",
                         "AlterColumnTypeEvent{tableId=%s.members, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}",
-                        "AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`precise_age` DOUBLE, position=LAST, existedColumnName=null}]}",
                         "AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`biological_sex` TINYINT, position=LAST, existedColumnName=null}]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013, Fiona, null, null, 16.0, null], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014, Gem, null, null, 17.0, null], op=INSERT, meta=()}"));
-
-        Assertions.assertThat(taskManagerConsumer.toString())
-                .doesNotContain(
-                        String.format(
-                                "Applied schema change event DropTableEvent{tableId=%s.members}",
-                                schemaEvolveDatabase.getDatabaseName()));
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013 -> Fiona, 1013, Fiona, 16.0, 1026169, age < 20, null, null], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014 -> Gem, 1014, Gem, 17.0, 1028196, age < 20, null, null], op=INSERT, meta=()}"));
     }
 
     @Test
@@ -188,105 +187,15 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                 true,
                 Arrays.asList(
                         "AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=age}]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012, Eve, 17, 0], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012 -> Eve, 1012, Eve, 17, 0, 1024144, age < 20], op=INSERT, meta=()}",
                         "AlterColumnTypeEvent{tableId=%s.members, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}",
-                        "RenameColumnEvent{tableId=%s.members, nameMapping={age=precise_age}}",
                         "RenameColumnEvent{tableId=%s.members, nameMapping={gender=biological_sex}}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013, Fiona, 16.0, null], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013 -> Fiona, 1013, Fiona, 16.0, null, 1026169, age < 20], op=INSERT, meta=()}",
                         "TruncateTableEvent{tableId=%s.members}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014, Gem, 17.0, null], op=INSERT, meta=()}"),
+                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014 -> Gem, 1014, Gem, 17.0, null, 1028196, age < 20], op=INSERT, meta=()}"),
                 Arrays.asList(
                         "Ignored schema change DropColumnEvent{tableId=%s.members, droppedColumnNames=[biological_sex]}.",
                         "Ignored schema change DropTableEvent{tableId=%s.members}."));
-    }
-
-    @Test
-    void testLenientWithRoute() throws Exception {
-        String dbName = schemaEvolveDatabase.getDatabaseName();
-
-        String pipelineJob =
-                String.format(
-                        "source:\n"
-                                + "  type: mysql\n"
-                                + "  hostname: %s\n"
-                                + "  port: 3306\n"
-                                + "  username: %s\n"
-                                + "  password: %s\n"
-                                + "  tables: %s.members\n"
-                                + "  server-id: 5400-5404\n"
-                                + "  server-time-zone: UTC\n"
-                                + "\n"
-                                + "route:\n"
-                                + "  - source-table: %s.members\n"
-                                + "    sink-table: %s.redirect\n"
-                                + "sink:\n"
-                                + "  type: values\n"
-                                + "\n"
-                                + "pipeline:\n"
-                                + "  schema.change.behavior: lenient\n"
-                                + "  parallelism: %d",
-                        INTER_CONTAINER_MYSQL_ALIAS,
-                        MYSQL_TEST_USER,
-                        MYSQL_TEST_PASSWORD,
-                        dbName,
-                        dbName,
-                        dbName,
-                        parallelism);
-        submitPipelineJob(pipelineJob);
-        waitUntilJobRunning(Duration.ofSeconds(30));
-        LOG.info("Pipeline job is running");
-        validateSnapshotData(dbName, "redirect");
-
-        LOG.info("Starting schema evolution");
-        String mysqlJdbcUrl =
-                String.format(
-                        "jdbc:mysql://%s:%s/%s", MYSQL.getHost(), MYSQL.getDatabasePort(), dbName);
-
-        try (Connection conn =
-                        DriverManager.getConnection(
-                                mysqlJdbcUrl, MYSQL_TEST_USER, MYSQL_TEST_PASSWORD);
-                Statement stmt = conn.createStatement()) {
-
-            waitForIncrementalStage(dbName, "redirect", stmt);
-
-            // triggers AddColumnEvent
-            stmt.execute("ALTER TABLE members ADD COLUMN gender TINYINT AFTER age;");
-            stmt.execute("INSERT INTO members VALUES (1012, 'Eve', 17, 0);");
-
-            // triggers AlterColumnTypeEvent and RenameColumnEvent
-            stmt.execute("ALTER TABLE members CHANGE COLUMN age precise_age DOUBLE;");
-
-            // triggers RenameColumnEvent
-            stmt.execute("ALTER TABLE members RENAME COLUMN gender TO biological_sex;");
-
-            // triggers DropColumnEvent
-            stmt.execute("ALTER TABLE members DROP COLUMN biological_sex");
-            stmt.execute("INSERT INTO members VALUES (1013, 'Fiona', 16);");
-
-            // triggers TruncateTableEvent
-            stmt.execute("TRUNCATE TABLE members;");
-            stmt.execute("INSERT INTO members VALUES (1014, 'Gem', 17);");
-
-            // triggers DropTableEvent
-            stmt.execute("DROP TABLE members;");
-        }
-
-        List<String> expectedTaskManagerEvents =
-                Arrays.asList(
-                        "AddColumnEvent{tableId=%s.redirect, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=LAST, existedColumnName=null}]}",
-                        "DataChangeEvent{tableId=%s.redirect, before=[], after=[1012, Eve, 17, 0], op=INSERT, meta=()}",
-                        "AlterColumnTypeEvent{tableId=%s.redirect, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}",
-                        "AddColumnEvent{tableId=%s.redirect, addedColumns=[ColumnWithPosition{column=`precise_age` DOUBLE, position=LAST, existedColumnName=null}]}",
-                        "AddColumnEvent{tableId=%s.redirect, addedColumns=[ColumnWithPosition{column=`biological_sex` TINYINT, position=LAST, existedColumnName=null}]}",
-                        "DataChangeEvent{tableId=%s.redirect, before=[], after=[1013, Fiona, null, null, 16.0, null], op=INSERT, meta=()}",
-                        "DataChangeEvent{tableId=%s.redirect, before=[], after=[1014, Gem, null, null, 17.0, null], op=INSERT, meta=()}");
-
-        String[] expectedTmEvents =
-                expectedTaskManagerEvents.stream()
-                        .map(s -> String.format(s, dbName, dbName))
-                        .toArray(String[]::new);
-
-        validateResult(expectedTmEvents);
     }
 
     @Test
@@ -318,99 +227,6 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
         // Submitting job should fail given an unknown schema change behavior configuration
         Assertions.assertThatThrownBy(() -> submitPipelineJob(pipelineJob))
                 .isExactlyInstanceOf(AssertionError.class);
-    }
-
-    @Test
-    void testByDefaultTransform() throws Exception {
-        String dbName = schemaEvolveDatabase.getDatabaseName();
-
-        // We put a dummy transform block that matches nothing
-        // to ensure TransformOperator exists, so we could verify if TransformOperator could
-        // correctly handle such "bypass" tables with schema changes.
-        String pipelineJob =
-                String.format(
-                        "source:\n"
-                                + "  type: mysql\n"
-                                + "  hostname: %s\n"
-                                + "  port: 3306\n"
-                                + "  username: %s\n"
-                                + "  password: %s\n"
-                                + "  tables: %s.members\n"
-                                + "  server-id: 5400-5404\n"
-                                + "  server-time-zone: UTC\n"
-                                + "\n"
-                                + "sink:\n"
-                                + "  type: values\n"
-                                + "transform:\n"
-                                + "  - source-table: another.irrelevant\n"
-                                + "    projection: \"'irrelevant' AS tag\"\n"
-                                + "\n"
-                                + "pipeline:\n"
-                                + "  schema.change.behavior: evolve\n"
-                                + "  parallelism: %d",
-                        INTER_CONTAINER_MYSQL_ALIAS,
-                        MYSQL_TEST_USER,
-                        MYSQL_TEST_PASSWORD,
-                        dbName,
-                        parallelism);
-        submitPipelineJob(pipelineJob);
-        waitUntilJobRunning(Duration.ofSeconds(30));
-        LOG.info("Pipeline job is running");
-        validateSnapshotData(dbName, "members");
-
-        LOG.info("Starting schema evolution");
-        String mysqlJdbcUrl =
-                String.format(
-                        "jdbc:mysql://%s:%s/%s", MYSQL.getHost(), MYSQL.getDatabasePort(), dbName);
-
-        try (Connection conn =
-                        DriverManager.getConnection(
-                                mysqlJdbcUrl, MYSQL_TEST_USER, MYSQL_TEST_PASSWORD);
-                Statement stmt = conn.createStatement()) {
-
-            waitForIncrementalStage(dbName, "members", stmt);
-
-            // triggers AddColumnEvent
-            stmt.execute("ALTER TABLE members ADD COLUMN gender TINYINT AFTER age;");
-            stmt.execute("INSERT INTO members VALUES (1012, 'Eve', 17, 0);");
-
-            // triggers AlterColumnTypeEvent and RenameColumnEvent
-            stmt.execute("ALTER TABLE members CHANGE COLUMN age precise_age DOUBLE;");
-
-            // triggers RenameColumnEvent
-            stmt.execute("ALTER TABLE members RENAME COLUMN gender TO biological_sex;");
-
-            // triggers DropColumnEvent
-            stmt.execute("ALTER TABLE members DROP COLUMN biological_sex");
-            stmt.execute("INSERT INTO members VALUES (1013, 'Fiona', 16);");
-
-            // triggers TruncateTableEvent
-            stmt.execute("TRUNCATE TABLE members;");
-            stmt.execute("INSERT INTO members VALUES (1014, 'Gem', 17);");
-
-            // triggers DropTableEvent
-            stmt.execute("DROP TABLE members;");
-        }
-
-        List<String> expectedTaskManagerEvents =
-                Arrays.asList(
-                        "AddColumnEvent{tableId=%s.members, addedColumns=[ColumnWithPosition{column=`gender` TINYINT, position=AFTER, existedColumnName=age}]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1012, Eve, 17, 0], op=INSERT, meta=()}",
-                        "AlterColumnTypeEvent{tableId=%s.members, typeMapping={age=DOUBLE}, oldTypeMapping={age=INT}, comments={}}",
-                        "RenameColumnEvent{tableId=%s.members, nameMapping={age=precise_age}}",
-                        "RenameColumnEvent{tableId=%s.members, nameMapping={gender=biological_sex}}",
-                        "DropColumnEvent{tableId=%s.members, droppedColumnNames=[biological_sex]}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1013, Fiona, 16.0], op=INSERT, meta=()}",
-                        "TruncateTableEvent{tableId=%s.members}",
-                        "DataChangeEvent{tableId=%s.members, before=[], after=[1014, Gem, 17.0], op=INSERT, meta=()}",
-                        "DropTableEvent{tableId=%s.members}");
-
-        String[] expectedTmEvents =
-                expectedTaskManagerEvents.stream()
-                        .map(s -> String.format(s, dbName, dbName))
-                        .toArray(String[]::new);
-
-        validateResult(expectedTmEvents);
     }
 
     private void testGenericSchemaEvolution(
@@ -457,7 +273,10 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                                 + (fineGrained
                                         ? "  exclude.schema.changes:\n" + "    - drop\n"
                                         : "")
-                                + (triggerError ? "  error.on.schema.change: true" : "")
+                                + (triggerError ? "  error.on.schema.change: true\n" : "\n")
+                                + "transform:\n"
+                                + "  - source-table: %s.\\.*\n"
+                                + "    projection: CAST(id AS VARCHAR) || ' -> ' || name AS uid, *, CASE WHEN age < 20 THEN id * id WHEN age >= 20 THEN 0 - id * id END AS id_square, CASE WHEN age < 20 THEN 'age < 20' WHEN age >= 20 THEN 'age >= 20' END as tag\n"
                                 + (mergeTable
                                         ? String.format(
                                                 "route:\n"
@@ -474,6 +293,7 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
                         MYSQL_TEST_PASSWORD,
                         dbName,
                         mergeTable ? "(members|new_members)" : "members",
+                        dbName,
                         behavior,
                         parallelism);
         submitPipelineJob(pipelineJob);
@@ -497,8 +317,8 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
             stmt.execute("ALTER TABLE members ADD COLUMN gender TINYINT AFTER age;");
             stmt.execute("INSERT INTO members VALUES (1012, 'Eve', 17, 0);");
 
-            // triggers AlterColumnTypeEvent and RenameColumnEvent
-            stmt.execute("ALTER TABLE members CHANGE COLUMN age precise_age DOUBLE;");
+            // triggers AlterColumnTypeEvent
+            stmt.execute("ALTER TABLE members MODIFY COLUMN age DOUBLE;");
 
             // triggers RenameColumnEvent
             stmt.execute("ALTER TABLE members RENAME COLUMN gender TO biological_sex;");
@@ -533,11 +353,11 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
     private void validateSnapshotData(String dbName, String tableName) throws Exception {
         validateResult(
                 s -> String.format(s, dbName, tableName),
-                "CreateTableEvent{tableId=%s.%s, schema=columns={`id` INT NOT NULL,`name` VARCHAR(17),`age` INT}, primaryKeys=id, options=()}",
-                "DataChangeEvent{tableId=%s.%s, before=[], after=[1008, Alice, 21], op=INSERT, meta=()}",
-                "DataChangeEvent{tableId=%s.%s, before=[], after=[1009, Bob, 20], op=INSERT, meta=()}",
-                "DataChangeEvent{tableId=%s.%s, before=[], after=[1010, Carol, 19], op=INSERT, meta=()}",
-                "DataChangeEvent{tableId=%s.%s, before=[], after=[1011, Derrida, 18], op=INSERT, meta=()}");
+                "CreateTableEvent{tableId=%s.%s, schema=columns={`uid` STRING,`id` INT NOT NULL,`name` VARCHAR(17),`age` INT,`id_square` INT,`tag` STRING}, primaryKeys=id, options=()}",
+                "DataChangeEvent{tableId=%s.%s, before=[], after=[1009 -> Bob, 1009, Bob, 20, -1018081, age >= 20], op=INSERT, meta=()}",
+                "DataChangeEvent{tableId=%s.%s, before=[], after=[1008 -> Alice, 1008, Alice, 21, -1016064, age >= 20], op=INSERT, meta=()}",
+                "DataChangeEvent{tableId=%s.%s, before=[], after=[1011 -> Derrida, 1011, Derrida, 18, 1022121, age < 20], op=INSERT, meta=()}",
+                "DataChangeEvent{tableId=%s.%s, before=[], after=[1010 -> Carol, 1010, Carol, 19, 1020100, age < 20], op=INSERT, meta=()}");
     }
 
     private void waitForIncrementalStage(String dbName, String tableName, Statement stmt)
@@ -548,7 +368,7 @@ class SchemaEvolveE2eITCase extends PipelineTestEnvironment {
         waitUntilSpecificEvent(
                 taskManagerConsumer,
                 String.format(
-                        "DataChangeEvent{tableId=%s.%s, before=[], after=[0, __fence__, 0], op=INSERT, meta=()}",
+                        "DataChangeEvent{tableId=%s.%s, before=[], after=[0 -> __fence__, 0, __fence__, 0, 0, age < 20], op=INSERT, meta=()}",
                         dbName, tableName));
     }
 }
